prod.requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.11
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
